backend/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dateutil==2.9.0.post0
psycopg[binary]==3.1.19